refactor(1064): move the EventsCard widget into Humans.Events
Shell owned a view component that reads Events data and localizes
EventsResource. It belongs to the section: component, view model and
Default.cshtml move across, the view model goes internal, and the tests
move to Humans.Events.Tests.

Refs nobodies-collective/Humans#1064
