+++ src/Humans.Web/Views/_ViewImports.cshtml
@@ -5,7 +5,6 @@
 @using Humans.Web.Extensions.Sections
 @using Humans.Web.Helpers
 @using Humans.Web.Models
-@using Humans.Web.Models.Events
 @using Humans.Web.ViewComponents
 @using Humans.UI.ViewComponents
 @using Humans.Domain.Enums
