+++ src/Sections/Humans.Events/Views/Shared/Components/EventsCard/Default.cshtml
@@ -1,8 +1,4 @@
-@model Humans.Web.Models.Events.EventsCardViewModel
-@* FavouriteButtonModel moved to Humans.Events at G5 lane 4b-i (nobodies-collective/Humans#866). *@
-@using Humans.Events.Contracts
-@* Events strings live in the section's resource set since #866 G5; Shell references the section project, so it localizes them directly. *@
-@inject IStringLocalizer<Humans.Events.EventsResource> EventsLocalizer
+@model EventsCardViewModel
 
 @functions {
     static string FormatDuration(int minutes)
@@ -38,10 +34,10 @@
                         {
                             EventId = row.EventId,
                             IsFavourited = row.IsFavourited,
-                            AddTitle = EventsLocalizer["Events_AddToFavourites"].Value,
-                            RemoveTitle = EventsLocalizer["Events_RemoveFromFavourites"].Value,
-                            RemoveRowTitle = EventsLocalizer["Events_RemoveFromSchedule"].Value,
-                            ErrorMessage = Localizer["Common_Error"].Value
+                            AddTitle = Localizer["Events_AddToFavourites"].Value,
+                            RemoveTitle = Localizer["Events_RemoveFromFavourites"].Value,
+                            RemoveRowTitle = Localizer["Events_RemoveFromSchedule"].Value,
+                            ErrorMessage = SharedLocalizer["Common_Error"].Value
                         })
                     </span>
                 </div>
