+++ tests/Humans.Events.Tests/EventsArchitectureTests.cs
@@ -75,23 +75,11 @@ public class EventsArchitectureTests
     [HumansFact]
     public void OnlySectionAndResourceArePublic()
     {
-        // "Public means Section or Contracts/" (design §15 step 5). Everything else in the
-        // assembly is internal, including the controllers: Shell registers
-        // SectionControllerFeatureProvider, which relaxes MVC's IsPublic check for assemblies
-        // carrying [assembly: Section("…")], so internal controllers still route
-        // (memory/architecture/section-controllers-need-feature-provider.md — which says in
-        // as many words: do not "fix" a 404 by making the controller public).
-        // The section's cross-section surface is mostly the separate Humans.Events.Contracts
-        // assembly. One exception since G5 lane 4b-i (nobodies-collective/Humans#866):
-        // FavouriteButtonModel, which moved here from Humans.UI and is bound by Shell's
-        // EventsCard component (src/Humans.Web/Views/Shared/Components/EventsCard/Default
-        // .cshtml) as well as this section's Browse and Schedule views. It is a view model,
-        // not a cross-section DTO, so it sits in this assembly's Contracts/ folder rather
-        // than in the leaf — the same placement Teams uses for HumansTeamControllerBase and
-        // Shifts for ShiftSignupsViewComponent.
-        // Generated migration classes are emitted `public partial` by `dotnet ef` and are
-        // never hand-edited (memory/process/never-hand-edit-migrations); Store's BaselineStore
-        // is public for the same reason. They are excluded rather than internalized.
+        // Public means Section, Contracts/, the resource marker, or a type the framework
+        // silently drops when internal. Controllers stay internal —
+        // SectionControllerFeatureProvider routes them; do not "fix" a 404 by going public.
+        // EventsCardViewComponent: Razor only builds <vc:events-card> from a public class.
+        // Migrations are emitted public by dotnet ef, so they are excluded below.
         var publicTypes = typeof(Section).Assembly.GetExportedTypes()
             .Where(t => !string.Equals(t.Namespace, "Humans.Events.Data.Migrations", StringComparison.Ordinal))
             .Select(t => t.FullName)
@@ -103,11 +91,10 @@ public class EventsArchitectureTests
                 "Humans.Events.Contracts.FavouriteButtonModel",
                 "Humans.Events.EventsResource",
                 "Humans.Events.Section",
+                "Humans.Events.ViewComponents.EventsCardViewComponent",
             ],
-            because: "a section exposes its ISection entry point, its resource marker and its "
-                   + "Contracts/ folder and nothing else; "
-                   + "the resource marker is public because the boot localization diagnostic "
-                   + "discovers it via GetExportedTypes()");
+            because: "a section exposes its entry point, resource marker, Contracts/ folder, "
+                   + "and what the framework needs public — nothing else");
     }
 
     [HumansFact]
